// src/helpers/joi/agent/account/account files/account_files.joi.types.ts
interface DeleteAppUserFileType {
  appAgentId: string | null;
  aadhar_cardId: string | null;
  pan_cardId: string | null;
  profile_pictureId: string | null;
  documentId: string | null;
}

interface GetAppUserFileType {
  appAgentId: string;
}

interface GetSingleFileType {
  attatchmentId: string;
}

interface File {
  fieldname: string;
  originalname: string;
  encoding: string;
  mimetype: string;
  destination: string;
  filename: string;
  path: string;
  size: number;
}

interface UpdatedFilesId {
  appAgentId: string;
  profile_pictureId: string | null;
  aadhar_cardId: string | null;
  pan_cardId: string | null;
  documentId: string | null;
}
interface UpdatedFiles {
  profile_picture: File[] | null;
  aadhar_card: File[] | null;
  pan_card: File[] | null;
  documents: File[] | null;
}
export { DeleteAppUserFileType, GetAppUserFileType, GetSingleFileType, UpdatedFilesId, UpdatedFiles };